feat: add connection panel screen
