+++ frontend/components/card.css
@@ -1,0 +1,43 @@
+.card {
+    background-color: #151515;
+    border-radius: 12px;
+    padding: 28px 25px;
+    box-shadow: 0 4px 20px rgba(0, 0, 0, 0.5);
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    gap: 8px;
+    user-select: none;
+}
+
+.card:hover {
+    background-color: #1d1d1d;
+    box-shadow: 0 4px 20px #1d1d1d;
+}
+
+.card h3 {
+    font-size: 0.97rem;
+    font-weight: 500;
+    text-transform: uppercase;
+    margin: 0;
+    line-height: 1;
+    color: #D4DADF;
+}
+
+.card .value {
+    font-size: 3.4rem;
+    font-weight: 700;
+    margin: 0;
+    line-height: 1;
+    color: #F7F3FF;
+}
+
+.card .unit {
+    font-size: 1.3rem;
+    font-weight: 200;
+    color: #D4DADF;
+}
+
+.card.dimmed * {
+    opacity: 0.15 !important;
+}
