+++ frontend/components/forms.css
@@ -1,0 +1,66 @@
+
+.vin-input::placeholder {
+    color: #49454F;
+    opacity: 1;
+}
+
+.search-container {
+    position: relative;
+    width: 50%;
+    display: flex;
+    align-items: center;
+    background-color: #151515;
+    border-radius: 10px;
+}
+
+.search {
+    position: relative;
+    font-size: .95rem;
+    background-color: #151515;
+    border-radius: 10px;
+    padding: 6px 20px;
+    width: 100%;
+    height: 35px;
+    border: none;
+    outline: none;
+    color: #F7F3FF;
+    box-shadow: 0 4px 20px rgba(0, 0, 0, 0.5);
+}
+
+.search-icon {
+    position: absolute;
+    right: 15px;
+    width: 16px;
+    height: 16px;
+    pointer-events: none;
+}
+
+.search::placeholder {
+    color: #49454F;
+    opacity: 1;
+}
+
+.vin-input-container {
+    position: relative;
+    width: 58%;
+    display: flex;
+    align-items: center;
+    background-color: #151515;
+    border-radius: 5px;
+}
+
+.vin-input {
+    position: relative;
+    font-size: .95rem;
+    background-color: #171717;
+    border-radius: 5px;
+    margin-top: 2px;
+
+    padding: 6px 20px;
+    width: 100%;
+    height: 35px;
+    border: none;
+    outline: none;
+    color: #F7F3FF;
+    box-shadow: 0 4px 10px rgba(0, 0, 0, 0.4);
+}
