+++ frontend/components/pid.css
@@ -1,0 +1,126 @@
+
+.pid-container:hover .pid-row,
+.pid-container:hover .pid-details {
+    background-color: #1d1d1d;
+    box-shadow: 0 4px 20px #1d1d1d;
+}
+
+.pid-row {
+    background-color: #151515;
+    height: 10px;
+    border-radius: 8px;
+    padding: 19px 10px;
+    display: flex;
+    align-items: center;
+    cursor: pointer;
+    margin-bottom: 8px;
+}
+
+.pid-row .name {
+    font-size: 1.0rem;
+    font-weight: 550;
+    padding-left: 10px;
+    color: #f3f3f2;
+}
+
+.pid-row.expanded {
+    border-bottom-left-radius: 0;
+    border-bottom-right-radius: 0;
+    box-shadow: none;
+}
+
+.pid-row.expanded .arrow-icon {
+    transform: rotate(-180deg);
+}
+
+.pid-row.expanded+.pid-details {
+    opacity: 1;
+    transform: translateY(0);
+}
+
+
+.pid-details {
+    height: 0;
+    opacity: 0;
+    overflow: hidden;
+    background-color: #151515;
+    margin-top: -12px;
+    margin-bottom: 10px;
+    transform: translateY(-10px);
+    box-shadow: 0 3px 10px rgba(0, 0, 0, 0.3);
+    border-radius: 0 0 12px 12px;
+    font-size: 0.95rem;
+    position: static;
+    padding: 0 12px;
+    transition: height 0.3s ease, opacity 0.3s ease, transform 0.3s ease;
+}
+
+.pid-data-columns {
+    display: flex;
+    gap: 40px;
+}
+
+.pid-column {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+}
+
+.pid-button {
+    margin-top: 12px;
+    margin-bottom: 10px;
+    height: 22px;
+    width: 85px;
+    background-color: #303030;
+    color: #F7F3FF;
+    font-weight: 700;
+    font-size: 0.8em;
+    border: none;
+    padding-left: 8px;
+    padding-right: 8px;
+    border-radius: 5px;
+    cursor: pointer;
+}
+
+.pid-label {
+    font-weight: 600;
+    font-size: 0.9rem;
+    color: #F7F3FF;
+    text-transform: uppercase;
+}
+
+.pid-value {
+    font-weight: 500;
+    font-size: 0.85rem;
+    color: #F7F3FF;
+    white-space: nowrap;
+}
+
+
+.pid-row-scroll {
+    overflow-y: auto;
+    flex-grow: 1;
+    padding-right: 10px;
+    padding-bottom: 50px;
+    height: 100%;
+    min-height: 0;
+}
+
+.pid-row-scroll::-webkit-scrollbar {
+    width: 13px;
+}
+
+.pid-row-scroll::-webkit-scrollbar-track {
+    background: #0F0F0F;
+    border-radius: 10px;
+}
+
+.pid-row-scroll::-webkit-scrollbar-thumb {
+    background-color: #29272c;
+    border-radius: 10px;
+    border: 2px solid #0F0F0F;
+}
+
+.pid-row-scroll::-webkit-scrollbar-thumb:hover {
+    background-color: #49454F;
+}
