+++ frontend/components/vin.css
@@ -1,0 +1,41 @@
+
+.vin {
+    font-size: 16px;
+    color: #BDBDBD;
+    margin-left: 1px;
+    font-weight: 500;
+}
+
+.car-image {
+    height: 215px;
+}
+
+.vin-results {
+    overflow-y: auto;
+    display: flex;
+    gap: 70px;
+    margin-top: 20px;
+    flex-wrap: wrap;
+    height: 100%;
+}
+
+.vin-column {
+    display: flex;
+    flex-direction: column;
+    gap: 16px;
+}
+
+.vin-label .label {
+    font-size: 16px;
+    color: #D4DADF;
+    font-weight: 500;
+    text-transform: uppercase;
+}
+
+.vin-label .value {
+    font-size: 20px;
+    font-weight: 700;
+    color: #F7F3FF;
+    margin-bottom: -7px;
+    max-width: 400px;
+}
