+++ frontend/index.html
@@ -7,6 +7,15 @@
   <title>OBDII Dashboard</title>
   <script type="module" src="main.js"></script>
   <link rel="stylesheet" href="styles.css">
+  <link rel="stylesheet" href="layout.css">
+  <link rel="stylesheet" href="scrollbar.css">
+  <link rel="stylesheet" href="components/card.css">
+  <link rel="stylesheet" href="components/forms.css">
+  <link rel="stylesheet" href="components/pid.css">
+  <link rel="stylesheet" href="components/sidebar.css">
+  <link rel="stylesheet" href="components/vin.css">
+  <link rel="stylesheet" href="components/buttons.css">
+
   <link href="https://fonts.googleapis.com/css2?family=Montserrat:wght@400;500;600;700&display=swap" rel="stylesheet">
 </head>
 
@@ -82,39 +91,81 @@
     </div>
     <div id="connect-settings" class="screen" style="display: none;">
       <div id="connect-header" class="screen-label">CONNECTION SETTINGS</div>
-      
+
+      <div class="connection-details" id="connection-details">
+        NO CONNECTION ESTABLISHED
+      </div>
+
       <div class="form-row">
-        <div class="dropdown" id="custom-dropdown">
+        <div class="dropdown" id="protocol-dropdown">
           <label>OBD-II PROTOCOL</label>
-          <button class="dropdown-toggle" id="dropdown-toggle"></button>
+          <button class="dropdown-toggle" id="dropdown-toggle">AUTOMATIC</button>
           <ul class="dropdown-menu" id="dropdown-menu">
-            <li data-value="option1">Option 1</li>
-            <li data-value="option2">Option 2</li>
-            <li data-value="option3">Option 3</li>
+            <li data-value="0">AUTOMATIC</li>
+            <li data-value="1">SAE J1850 PWM @41.6k</li>
+            <li data-value="2">SAE J1850 VPW @10.4k</li>
+            <li data-value="3">ISO 9141-2 (5 baud init)</li>
+            <li data-value="4">ISO 14230-4 KWP (5 baud init)</li>
+            <li data-value="5">ISO 14230-4 KWP (fast init)</li>
+            <li data-value="6">ISO 15765-4 CAN (11 bit ID @500k)</li>
+            <li data-value="7">ISO 15765-4 CAN (29 bit ID @500k)</li>
+            <li data-value="8">ISO 15765-4 CAN (11 bit ID @250k)</li>
+            <li data-value="9">ISO 15765-4 CAN (29 bit ID @250k)</li>
           </ul>
         </div>
-  
-        <div class="dropdown" id="custom-dropdown">
-          <label>SERIAL PORT COM</label>
-          <button class="dropdown-toggle" id="dropdown-toggle"></button>
-          <ul class="dropdown-menu" id="dropdown-menu">
-            <li data-value="option1">Option 1</li>
-            <li data-value="option2">s 2</li>
-            <li data-value="option3">Option 3</li>
-          </ul>
+
+        <div class="dropdown" id="serial-port-dropdown">
+          <div class="label-row">
+            <label>SERIAL PORT COM</label>
+            <button class="refresh" id="refresh-serial-ports">
+              <svg xmlns="http://www.w3.org/2000/svg" width="17" height="17" viewBox="0 0 24 24" fill="none"
+                stroke="#f3f3f2" stroke-width="2.5" stroke-linecap="round" stroke-linejoin="round"
+                class="lucide lucide-refresh-cw-icon lucide-refresh-cw">
+                <path d="M3 12a9 9 0 0 1 9-9 9.75 9.75 0 0 1 6.74 2.74L21 8" />
+                <path d="M21 3v5h-5" />
+                <path d="M21 12a9 9 0 0 1-9 9 9.75 9.75 0 0 1-6.74-2.74L3 16" />
+                <path d="M8 16H3v5" />
+              </svg>
+            </button>
+          </div>
+
+          <button class="dropdown-toggle" id="serial-port-selected">NO PORTS FOUND</button>
+          <ul class="dropdown-menu" id="serial-port-dropdown-menu"></ul>
         </div>
-  
-        <div class="dropdown" id="custom-dropdown">
+
+        <div class="dropdown" id="baud-rate-dropdown">
           <label>BAUD RATE</label>
-          <button class="dropdown-toggle" id="dropdown-toggle"></button>
+          <button class="dropdown-toggle" id="baud-rate-selected">38400</button>
           <ul class="dropdown-menu" id="dropdown-menu">
-            <li data-value="option1">Option 1</li>
-            <li data-value="option2">Option 2</li>
-            <li data-value="option3">x 3</li>
+            <li data-value="4800">4800</li>
+            <li data-value="9600">9600</li>
+            <li data-value="19200">19200</li>
+            <li data-value="38400">38400</li>
+            <li data-value="57600">57600</li>
+            <li data-value="115200">115200</li>
           </ul>
         </div>
       </div>
-      
+
+      <div class="button-row">
+        <button class="btn connect" id="btn-connect">
+          CONNECT
+        </button>
+        <button class="btn disconnect" id="btn-disconnect" disabled>
+          DISCONNECT
+        </button>
+      </div>
+
+      <div class="error-hint" style="display: flex;">
+        <svg xmlns="http://www.w3.org/2000/svg" width="15" height="15" viewBox="0 0 24 24" fill="none"
+          stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round">
+          <circle cx="12" cy="12" r="10" />
+          <path d="M12 16v-4" />
+          <path d="M12 8h.01" />
+        </svg>
+        CHOOSE SERIAL PORT, BAUD RATE AND PROTOCOL
+      </div>
+
     </div>
     <div id="dashboard" class="screen" style="display: flex;">
       <div id="dashboard-header" class="screen-label">OBD OVERVIEW</div>
@@ -319,7 +370,7 @@
   </div>
 </body>
 
-<script>
+<script>8
   const { listen, emit } = window.__TAURI__.event;
 
   /// Search bar functionality for dashboard
@@ -442,6 +493,21 @@
       card.classList.add('dimmed');
       dashboard.appendChild(card);
     }
+  });
+
+  /// Refresh serial ports
+
+  const refreshButton = document.getElementById('refresh-serial-ports');
+  const refreshIcon = refreshButton.querySelector('svg');
+
+  refreshButton.addEventListener('click', () => {
+    refreshIcon.classList.add('spinning');
+
+    emit('get-serial-ports');
+  });
+
+  refreshIcon.addEventListener('animationend', () => {
+    refreshIcon.classList.remove('spinning');
   });
 </script>
