+++ frontend/scrollbar.css
@@ -1,0 +1,38 @@
+.grid::-webkit-scrollbar {
+    width: 13px;
+}
+
+.grid::-webkit-scrollbar-track {
+    background: #0F0F0F;
+    border-radius: 10px;
+}
+
+.grid::-webkit-scrollbar-thumb {
+    background-color: #29272c;
+    border-radius: 10px;
+    border: 2px solid #0F0F0F;
+}
+
+.grid::-webkit-scrollbar-thumb:hover {
+    background-color: #49454F;
+}
+
+
+.vin-results::-webkit-scrollbar {
+    width: 13px;
+}
+
+.vin-results::-webkit-scrollbar-track {
+    background: #0F0F0F;
+    border-radius: 10px;
+}
+
+.vin-results::-webkit-scrollbar-thumb {
+    background-color: #29272c;
+    border-radius: 10px;
+    border: 2px solid #0F0F0F;
+}
+
+.vin-results::-webkit-scrollbar-thumb:hover {
+    background-color: #49454F;
+}
